Refactor and use different npm package for dropzone and drag/move image
