+++ package.json
@@ -13,12 +13,10 @@
     "material-ui-dropzone": "^3.2.1",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
+    "react-avatar-editor": "^11.0.9",
     "react-color-extractor": "^1.1.2",
-    "react-dnd": "^11.1.3",
-    "react-dnd-html5-backend": "^11.1.3",
     "react-dom": "^16.13.1",
     "react-dropzone": "^11.0.1",
-    "react-image-file-resizer": "^0.3.1",
     "react-is": "^16.13.1",
     "react-redux": "^7.2.0",
     "react-router": "^5.2.0",
