--- src/components/DragNDrop/Box.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const styles = {
-  padding: "0 0",
-  cursor: "move",
-  height: "100%",
-  width: "auto",
-};
-export const Box = ({ img, yellow }) => {
-  const backgroundColor = yellow ? "yellow" : "white";
-  return <img src={img} style={{ ...styles, backgroundColor }}/>;
-};
