--- src/components/DragNDrop/CustomDragLayer.js
@@ -1,61 +1,0 @@
-import React from "react";
-import { DragLayer } from "react-dnd";
-import { ItemTypes } from "./ItemTypes";
-import { BoxDragPreview } from "./BoxDragPreview";
-import { snapToGrid } from "./snapToGrid";
-const layerStyles = {
-  position: "fixed",
-  pointerEvents: "none",
-  zIndex: 100,
-  left: 0,
-  top: 0,
-  width: "100%",
-  height: "100%",
-};
-function getItemStyles(props) {
-  const { initialOffset, currentOffset } = props;
-  if (!initialOffset || !currentOffset) {
-    return {
-      display: "none",
-    };
-  }
-  let { x, y } = currentOffset;
-  if (props.snapToGrid) {
-    x -= initialOffset.x;
-    y -= initialOffset.y;
-    [x, y] = snapToGrid(x, y);
-    x += initialOffset.x;
-    y += initialOffset.y;
-  }
-  const transform = `translate(${x}px, ${y}px)`;
-  return {
-    transform,
-    WebkitTransform: transform,
-  };
-}
-const CustomDragLayer = (props) => {
-  const { item, img, itemType, isDragging } = props;
-  function renderItem() {
-    switch (itemType) {
-      case ItemTypes.BOX:
-        return <BoxDragPreview img={img} />;
-      default:
-        return null;
-    }
-  }
-  if (!isDragging) {
-    return null;
-  }
-  return (
-    <div style={layerStyles}>
-      <div style={getItemStyles(props)}>{renderItem()}</div>
-    </div>
-  );
-};
-export default DragLayer((monitor) => ({
-  item: monitor.getItem(),
-  itemType: monitor.getItemType(),
-  initialOffset: monitor.getInitialSourceClientOffset(),
-  currentOffset: monitor.getSourceClientOffset(),
-  isDragging: monitor.isDragging(),
-}))(CustomDragLayer);
