--- src/components/DragNDrop/ItemTypes.js
@@ -1,3 +1,0 @@
-export const ItemTypes = {
-  BOX: "box",
-};
