+++ src/components/SignInPage.js
@@ -2,17 +2,14 @@ import React from "react";
 import Avatar from "@material-ui/core/Avatar";
 import Button from "@material-ui/core/Button";
 import CssBaseline from "@material-ui/core/CssBaseline";
-import TextField from "@material-ui/core/TextField";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
-import Link from "@material-ui/core/Link";
 import Paper from "@material-ui/core/Paper";
-import Box from "@material-ui/core/Box";
 import Grid from "@material-ui/core/Grid";
 import LockOutlinedIcon from "@material-ui/icons/LockOutlined";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
-import DragNDrop from "./DragNDrop/DragNDrop";
+import DragNDrop from "./DragNDrop";
 import EmailInput from "./InputForm/EmailInput";
 import PasswordInput from "./InputForm/PasswordInput";
 import ForgotPassword from "./InputForm/ForgotPassword";
@@ -65,7 +62,15 @@ export default function SignInPage({
   return (
     <Grid container component="main" className={classes.root}>
       <CssBaseline />
-      <Grid item xs={false} sm={4} md={6} lg={8} xl={9}>
+      <Grid
+        item
+        xs={false}
+        sm={4}
+        md={6}
+        lg={8}
+        xl={9}
+        style={{ maxHeight: "100%" }}
+      >
         <DragNDrop />
       </Grid>
 
@@ -79,6 +84,7 @@ export default function SignInPage({
         component={Paper}
         elevation={6}
         square
+        style={{ maxHeight: "100%" }}
       >
         <div className={classes.paper}>
           <Avatar className={classes.avatar}>
@@ -112,11 +118,10 @@ export default function SignInPage({
             <Grid
               item
               container
-              disableGutters
-              maxWidth={false}
+              maxwidth="false"
               direction="column"
               justify="flex-end"
-              style={{ height: "45vh" }}
+              style={{ height: "30vh", maxHeight: "30vh" }}
             >
               <Grid item zeroMinWidth>
                 <Copyright />
